+++ src/games/Trivia.py
@@ -4,39 +4,10 @@ import sqlalchemy
 from urllib.parse import unquote
 from random import shuffle
 
-class Game:
-  def __init__(self, engine, key, data):
-    self.stages = {}
-    self.players = {}
-    self.max_players = 8
-    self.data = data
-    self.engine = engine
-    self.currentStage = 0
-    self.key = key
-
-    self.start()
-
-  def add_player(self, username):
-    if self.currentStage == 0 and len(self.players) < self.max_players:
-      self.players[username] = True
-      return True
-    return False
-
-  def start(self):
-    return None
-
-  def end(self):
-    return None
-
-  @property
-  def game_state(self):
-    return {
-      "players": len(self.players),
-      "key": self.key,
-    }
+from .GameHandler import GameHandler
 
 
-class Trivia(Game):
+class Trivia(GameHandler):
   def start(self):
     self.stages = {}
     for i in range(int(self.data["numRounds"])):
