+++ src/hackbox.py
@@ -6,18 +6,20 @@
 """
 
 from flask import Flask, render_template, jsonify, request, session, redirect, send_from_directory
-from config import Config
-from login import register_user, login_user, message_format, dest_format, logged_in
-from Trivia import Trivia
-from HackerFall import HackerFall
-from get_key import get_key
 from sql import engine
+
+from modules.config import Config
+from modules.login import register_user, login_user, message_format, dest_format, logged_in
+from modules.get_key import get_key
+from games.Trivia import Trivia
+from games.HackerFall import HackerFall
+
 app = Flask(__name__)
 
 ### possible games
 games = {
-    'Trivia':Trivia,
-    "HackerFall":HackerFall,
+    'Trivia': Trivia,
+    "HackerFall": HackerFall,
 }
 current_games = {}
